Creating Documentation folder in project root. (#114)
PR #8 

This PR is dedicated to adding the application documentation to the
project.

1) This PR implements the changes mentioned. I have not moved docs/
folder to project root, as docs/ already contains the OPENAPI
Documentation code, which is integrated into the app’s source code and
actively used by the application.

2) A new documentation/ folder has been created in the project root.
This folder contains detailed documentation for the app.

I would appreciate it if you could review the changes and let me know if
any modifications are needed. If any changes or adjustments are
required, I will address them promptly. Thankyou!

---------

Co-authored-by: codecakes <codecakes@users.noreply.github.com>